Fix search, overall
There were a couple issues with the search mechanism:
* Almost all components down the search line would save and duplicate
some part of the search state.
* This lead to race conditions and state mismatches between components,
so it would be impossible to change a search query without reloading the
page completely.

Basically, this [*lifts the state up*][1] all the way to the
WrappedDataset component.

WrappedDataset hold the query state, Dataset pushes updates to
WrappedDataset’s state and SearchInput’s input is no longer a controlled
input.

I also changed the SearchInput to use a form instead of keyUp hooks for
submitting, as it’s how it’s done in HTML.

[1]: https://facebook.github.io/react/docs/lifting-state-up.html