Pull CI reports from current branch or master only
Stop pulling reports from random branches if there is no reports cache
for the current branch as the diff would be irrelevant.

We either take the reports from the last build on the branch or the last
build on master. If there is none available (which will happen only if
the cache is cleared), we won’t be displaying the diff.

Also remove the over specific branch-commit pull as it can never happen.