Make CenteredMap component loadable asynchronously
This moves all leaflet into an async component, so it won’t be bundled
in vendors anymore. It removes about 56KB from the vendor bundle.

On the dataset page, the preview will now show a spinner while the
bundle is loading (the first time a user gets to a dataset page).