Stop serving leaflet css from cloudflare
Add vendor css rule in webpack that inlines CSS in the JS modules. It
has some overhead but, it’s currently the only way for the styles to be
delivered asynchronously.

The sourcemaps are disabled in production as they are always inline (and
not in their own .map file). Since we do not control those files, the
results will always be the same in development, so development
sourcemaps is enough (and that’s a 5KB gzipped gain).

Reference #413