Make sure footer is sticking to the bottom
The previous behaviour would make the content of the page (excluding
header and footer) 100vh minimum. The issue with that is that we would
always have a scroll on small pages of the size of the header.

This allows to have the footer visible on pages with little content.

`flex: 1` allows the content of the page to grow, hence push the footer
to the bottom.